ir module
